\name{print.twiddlerControl}
\alias{print.twiddlerControl}
\title{Display information about a twiddler control...}
\usage{\method{print}{twiddlerControl}(x, ...)
}
\description{Display information about a twiddler control}
\value{Invisibly returns \code{x}.}
\arguments{\item{x}{A twiddler control}
\item{...}{Ignored}
}

