\name{print.twiddlerControl}
\alias{print.twiddlerControl}
\title{Display information about a twiddler control}
\usage{
  \method{print}{twiddlerControl} (x, ...)
}
\arguments{
  \item{x}{A twiddler control}

  \item{...}{Ignored}
}
\value{
  Invisibly returns \code{x}.
}
\description{
  Display information about a twiddler control
}

