\name{shortenString}
\alias{shortenString}
\title{Shorten a string if it is longer than a limit}
\usage{
  shortenString(text, limit, cutIndication = "...")
}
\arguments{
  \item{text}{The string to shorten.}

  \item{limit}{The length limit.}

  \item{cutIndication}{A string to append to the string if
  shortened.}
}
\value{
  The shortened string.
}
\description{
  Shorten a string if it is longer than a limit
}

