\name{twiddle_print}
\alias{twiddle_print}
\title{Print, then directly flush the output buffer...}
\usage{twiddle_print(...)
}
\description{Print, then directly flush the output buffer}
\details{This function can be used to "print in realtime" on platforms that buffer console output,
like on Windows.}
\arguments{\item{...}{Passed on to \code{\link{print}}}
}

