\name{toggle}
\alias{toggle}
\title{Checkbox to manipulate a boolean valued variable...}
\usage{toggle(default=FALSE, label=as.character(NA))
}
\description{Checkbox to manipulate a boolean valued variable}
\details{\code{toggle} creates a checkbox \code{twiddlerControl} for manipulating
logical variables.}
\value{An object of class \code{twiddlerControl}.}
\seealso{\code{\link{twiddle}}, \code{\link{combo}}, \code{\link{entry}},
and \code{\link{knob}}}
\arguments{\item{default}{The default logical value for the checkbox, \code{TRUE}
means "checked", \code{FALSE} "unchecked".}
\item{label}{The text label of the checkbox.}
}

